add trait for add / update and fix updatePost previous image deleted
